Split bootstrap.inc.py for better debugging
