+++ pythonscript/embedded/godot/hazmat/ffi/__init__.py
@@ -1,0 +1,5 @@
+from godot.hazmat.ffi.init import *  # noqa
+from godot.hazmat.ffi.profiler import *  # noqa
+from godot.hazmat.ffi.editor import *  # noqa
+from godot.hazmat.ffi.script import *  # noqa
+from godot.hazmat.ffi.instance import *  # noqa
